+++ modules/aws-load-balancer-controller/variables.tf
@@ -1,0 +1,24 @@
+variable "cluster_name" {
+  type        = string
+  description = "Name of the EKS cluster"
+}
+
+variable "vpc_id" {
+  type        = string
+  description = "VPC ID where the cluster is deployed"
+}
+
+variable "region" {
+  type        = string
+  description = "AWS region"
+}
+
+variable "oidc_provider_arn" {
+  type        = string
+  description = "ARN of the EKS OIDC provider"
+}
+
+variable "oidc_provider" {
+  type        = string
+  description = "OIDC provider URL without https://"
+}